pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "vislang-ultralow-resource"
version = "0.1.0"
description = "Dataset builder and training framework for visual-language models in ultra-low-resource languages"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "Daniel Schmidt"},
]
maintainers = [
    {name = "Daniel Schmidt"},
]
keywords = [
    "computer-vision",
    "natural-language-processing", 
    "multilingual",
    "humanitarian",
    "low-resource-languages",
    "vision-language-models"
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
requires-python = ">=3.8"
dependencies = [
    "torch>=1.12.0",
    "transformers>=4.25.0",
    "datasets>=2.8.0",
    "Pillow>=9.0.0",
    "requests>=2.28.0",
    "beautifulsoup4>=4.11.0",
    "pandas>=1.5.0",
    "numpy>=1.21.0",
    "tqdm>=4.64.0",
    "sqlalchemy>=1.4.0",
    "psycopg2-binary>=2.9.0",
    "redis>=4.0.0",
    "click>=8.0.0",
    "fastapi>=0.88.0",
    "uvicorn[standard]>=0.20.0",
    "python-multipart>=0.0.5",
    "opencv-python>=4.6.0",
    "PyMuPDF>=1.20.0",
    "scikit-learn>=1.1.0",
    "accelerate>=0.15.0",
    "evaluate>=0.4.0",
]

[project.optional-dependencies]
ocr = [
    "pytesseract>=0.3.10",
    "easyocr>=1.6.0",
    "paddleocr>=2.6.0",
]
training = [
    "accelerate>=0.15.0",
    "wandb>=0.13.0",
    "tensorboard>=2.11.0",
    "evaluate>=0.4.0",
]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "pytest-xdist>=3.0.0",
    "black>=22.0.0",
    "isort>=5.10.0",
    "flake8>=5.0.0",
    "mypy>=0.991",
    "pre-commit>=2.20.0",
    "jupyterlab>=3.5.0",
    "ipywidgets>=8.0.0",
    "nbformat>=5.0.0",
    "sphinx>=5.0.0",
    "sphinx-rtd-theme>=1.0.0",
    "myst-parser>=0.18.0",
]
all = [
    "vislang-ultralow-resource[ocr,training,dev]"
]

[project.scripts]
vislang = "vislang_ultralow.cli:cli"

[project.urls]
Homepage = "https://github.com/danieleschmidt/quantum-inspired-task-planner"
Documentation = "https://github.com/danieleschmidt/quantum-inspired-task-planner#readme"
Repository = "https://github.com/danieleschmidt/quantum-inspired-task-planner.git"
"Bug Tracker" = "https://github.com/danieleschmidt/quantum-inspired-task-planner/issues"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-dir]
"" = "src"

[tool.black]
line-length = 88
target-version = ['py38']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 88
known_first_party = ["vislang_ultralow"]

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "--cov=src --cov-report=html --cov-report=term-missing --cov-fail-under=80"

[tool.coverage.run]
source = ["src"]
omit = [
    "*/tests/*",
    "*/test_*",
    "*/__pycache__/*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
]